+++ react/postalCodeFrom/TwoLevels.js
@@ -60,7 +60,7 @@ class TwoLevels extends Component {
 
 TwoLevels.propTypes = {
   address: AddressShapeWithValidation,
-  Button: PropTypes.func.isRequired,
+  Button: PropTypes.func,
   Input: PropTypes.func.isRequired,
   onChangeAddress: PropTypes.func.isRequired,
   onSubmit: PropTypes.func,